Fixed hot reloading with net code + minor changes
When hot reloading is enabled then game.c and net.c get compiled into
game.so, but because net.c starts new threads, the threads still run the
old code, which causes a segfault when the game.so is reloaded.

To fix this, I made the net code compile into a separate shared library,
which is loaded by the game code. Another solution could have been make
net.c compile with main.c and pass the function pointers of the net code
to the game code, but the chosen solution would make it easier to add
hot reloading of the net code in the future.